% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Get the mean std and se from a set of data with different variables}
\usage{
data_summary(in_data, varname, groupnames)
}
\arguments{
\item{in_data}{dataframe to be passed in (long format)}

\item{varname}{the variable to be tested}

\item{group_names}{groupnames of the data}
}
\value{
Function returns matrix.
}
\description{
Get the mean std and se from a set of data with different variables
}
