% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLS_autoplot.R
\name{WLS_autoplot}
\alias{WLS_autoplot}
\title{Plot WLS or return data by worm for .ods files}
\usage{
WLS_autoplot(
  lifespan_type = "RNAi",
  plot_individual_lifespans_to_control = TRUE,
  lifespan_directories = list.dirs()[list.dirs() != "."],
  return_data = FALSE,
  writeout_data = FALSE,
  empty_vector = c("L4440", "PAD12"),
  Vehicle = c("DMSO", "Vehicle"),
  Wild_type = c("Wild type", "Wild-type", "Wildtype", "wild type", "wildtype",
    "wild-type", "N2"),
  col_list = append(c("black", "red"), brewer.pal(8, "Set1")[3:8]),
  document_researchers = TRUE,
  col_names = c("Dead", "M-E", "M-B", "M", "M_all", "Day"),
  omit_file = "stack",
  processed_dir_path = FALSE,
  researchers = as.character(c()),
  Temperature = TRUE,
  lifespanfile_convert = FALSE,
  censored_plate = "x"
)
}
\arguments{
\item{lifespan_type}{This is where you put in what the lifespan type is. Accepted inputs: "RNAi", "Genotype", "Drug", "Treatment"}

\item{plot_individual_lifespans_to_control}{If TRUE, the individual lifespan plots compared to the control will be made.}

\item{lifespan_directories}{Directory in which will have all directories that are needed to be processed.}

\item{return_data}{If TRUE, this will return the data used for instead of making a graph survival plotting. This function cannot be done in batch processing files and will stop at the first directory}

\item{writeout_data}{#If TRUE, the data will be written out as a csv.}

\item{empty_vector}{If RNAi experiment, this is the empty vector.}

\item{Vehicle}{If drug experiment, what was the control, which would be used for stats.}

\item{Wild_type}{If Genotype experiment, what was the control, which would be used for stats description}

\item{col_list}{This controls the color of the lifespan plot lines.}

\item{document_researchers}{If TRUE, it will put the researchers names at the bottom of the plot.}

\item{col_names}{The names of the columns of the summary portion of the spreadsheet.}

\item{omit_file}{Pattern of file to omit in the analysis. The McCormick lab uses an ods file with "stack" in the filename to describe the experiment.}

\item{processed_dir_path}{Pass in a processed file if you'd like to skip the processing from the raw data collection spreadsheet provided}

\item{researchers}{default is to start with an empty vector as researchers.}

\item{Temperature}{If TRUE, the temperature will be provided in the plot.}

\item{lifespanfile_convert}{if TRUE this will convert the ods files to csv files for WLS_autoplot()}

\item{censored_plate}{What the censored plate code is. 'x' by default.}
}
\value{
Function can return the data that is used for lifespan plotting, or it saves lifespan plots out to the working directory.
}
\description{
Plot WLS or return data by worm for .ods files
}
