% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLS_convert_counts.R
\name{WLS_convert_counts}
\alias{WLS_convert_counts}
\title{Plot WLS manually}
\usage{
WLS_convert_counts(
  lifespan_type = "Genotype",
  counts_file = "counts.csv",
  rosetta_stone_file = "rosetta_stone.csv",
  rosettastone_index_colnumber = 1,
  lifespancounts_index_colnumber = 1,
  rosetta_stone_platenumber_replacement = "",
  lifespan_counts_platenumber_replacement = "",
  lifespancounts_count_colnumber = 6,
  convert_from_epoch_time = TRUE,
  processed_dir_path = FALSE,
  empty_vector = "L4440",
  Vehicle = "Vehicle",
  Wild_type = "N2"
)
}
\arguments{
\item{rosettastone_index_colnumber}{column number of the plate annotation in the rosetta_stone_file}

\item{lifespancounts_index_colnumber}{column number of the plate annotation in the counts_file}

\item{rosetta_stone_platenumber_replacement}{if there is a string to delete in the rosetta_stone file for plate column, put here}

\item{lifespan_counts_platenumber_replacement}{if there is a string to delete in the lifespan_counts file for plate column, put here}

\item{lifespancounts_count_colnumber}{colnumber where the counts are recorded in the lifespan_counts, if the 'Count" column is not labeled}

\item{convert_from_epoch_time}{The timestamp is defaul epoch time in the counts_file}

\item{processed_dir_path}{path to save the processed file}

\item{empty_vector}{If RNAi experiment, this is the empty vector.}

\item{Vehicle}{If drug experiment, what was the control, which would be used for stats.}

\item{Wild_type}{If Genotype experiment, what was the control, which would be used for stats description}

\item{lifespan_counts}{A matrix with the lifespan data}
}
\value{
Lifespan data to be plotted
}
\description{
Plot WLS manually
}
