% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLS_manualplot.R
\name{WLS_manualplot}
\alias{WLS_manualplot}
\title{Plot WLS manually}
\usage{
WLS_manualplot(
  lifespan_data,
  lifespan_type,
  col_list,
  plot_title,
  researchers = c(),
  x_upper_bound = 25,
  plot_individual_lifespans_to_control = FALSE,
  stats = FALSE,
  control = "",
  temperature = NA
)
}
\arguments{
\item{lifespan_data}{A matrix with the lifespan data}

\item{lifespan_type}{The type of lifespan experiment. Accepted: "RNAi", "Treatment", "Drug", "Genotype"}

\item{col_list}{color list to be used in plotting the lifespan}

\item{plot_title}{Title of the plot}

\item{researchers}{who the researchers are for the experiment. default = c()}

\item{x_upper_bound}{The x upper limit of the plot}

\item{plot_individual_lifespans_to_control}{If TRUE and stats = TRUE, write-out individual lifespans in comparison to the input control}

\item{stats}{If true, the function will plot the lifespan data with appropriate wilcoxon stats if given the correct label control in the lifespan_data}

\item{control}{The comparison used as a control if stats == TRUE}

\item{temperature}{input temperature value}
}
\value{
Lifespan plot with manual inputs
}
\description{
Plot WLS manually
}
